% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmapLFMC.R
\name{heatmapLFMC}
\alias{heatmapLFMC}
\title{Heatmap plots for LFMC data}
\usage{
heatmapLFMC(speciesCode = 1, siteCode = 1, period = "Fortnight")
}
\arguments{
\item{speciesCode}{Integer indicating the species code.}

\item{siteCode}{Integer indicating the site code.}

\item{period}{String indicating if values are shown on a monthly or a fortnightly basis.}
}
\value{
A heatmap by site
}
\description{
Heatmap plots for LFMC data
}
\examples{
\dontrun{
   heatmapLFMC(variable = "LFMC", speciescode = 5, sitecode = 50, period = "Month")
}

}
