% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_DB.R
\name{init_DB}
\alias{init_DB}
\title{Initializes an empty LFMC database with the appropriate database structure on a new file}
\usage{
init_DB(file, overwrite = FALSE)
}
\arguments{
\item{file}{Database filename}

\item{overwrite}{Boolean flag to force overwritting an existing file}
}
\description{
Initializes an empty LFMC database with the appropriate database structure on a new file
}
\examples{

init_DB("../lfmc")

}
