% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initDB.R
\name{initDB}
\alias{initDB}
\title{Initialize LFMC database}
\usage{
initDB(file, overwrite = FALSE)
}
\arguments{
\item{file}{Database filename}

\item{overwrite}{Boolean flag to force overwritting an existing file}
}
\description{
Initializes an empty LFMC database with the appropriate database structure on a new file
}
\examples{

\dontrun{
  init_DB("../lfmc", overwrite=TRUE)
}

}
