% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_LFMC.R
\name{parse_LFMC}
\alias{parse_LFMC}
\title{Reads LFMC data from an excel file}
\usage{
parse_LFMC(lfmc_xlsx, dateIni = NULL, dateFin = NULL,
  dateFormat = "\%d/\%m/\%Y", varmapping = c(Date = "DATA", SiteCode =
  "CODI_PARCELA", SpeciesCAT = "ESPECIE", SampleCode = "NUM_MOSTRA",
  FreshMass = "PES_FRESC", DryMass = "PES_SEC", LFMC = "HUMITAT", DryStem =
  "PES_TIGES", DryLeaf = "PES_FULLES", Notes = "Observacions"),
  overwrite = FALSE)
}
\arguments{
\item{overwrite}{}
}
\description{
Reads LFMC data from an excel file
}
