% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_DBfile.R
\name{set_DBpath}
\alias{set_DBpath}
\title{Specifies the LFMC database file}
\usage{
set_DBpath(file)
}
\arguments{
\item{file}{A string with the path to a *.sqlite file}
}
\description{
Specifies the LFMC database file
}
