% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmapLFMC.R
\name{heatmapLFMC}
\alias{heatmapLFMC}
\title{Heatmap plots for LFMC data}
\usage{
heatmapLFMC(
  variable = "LFMC",
  speciesCode = 1,
  siteCode = 1,
  period = "Fortnight"
)
}
\arguments{
\item{variable}{String indicating the variable to be graphed.}

\item{speciesCode}{Integer indicating the species code.}

\item{siteCode}{Integer or list of values indicating site codes.}

\item{period}{String indicating if LFMC values are shown on a monthly or a fortnightly basis.}
}
\value{
A heatmap by site
}
\description{
Heatmap plots for LFMC data
}
\examples{
\dontrun{
   heatmapLFMC(variable = "LFMC", speciesCode = 5, siteCode = 50, period = "Month")
}

}
