% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_DB.R
\name{init_DB}
\alias{init_DB}
\title{Initialize LFMC database}
\usage{
init_DB(file, thesaurus_xlsx = NULL, overwrite = FALSE)
}
\arguments{
\item{file}{Database filename}

\item{thesaurus_xlsx}{Excel 'xlsx' file with thesaurus tables for 'plots',
'sites', 'species' and 'agents'.}

\item{overwrite}{Boolean flag to force overwritting an existing file}
}
\description{
Initializes an empty LFMC database with the appropriate database structure on a new file
}
\examples{

\dontrun{
  init_DB("../lfmc", thesaurus_xlsx = "../LFMC_tesaures.xlsx")
}


}
