% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonalPlots.R
\name{seasonalPlot}
\alias{seasonalPlot}
\title{Seasonal plots with quantiles for LFMC data}
\usage{
seasonalPlot(
  speciescode = 1,
  sitecode = 1,
  period = "Fortnight",
  quantiles = c(0.1, 0.5, 0.95),
  LFMCflag = T,
  AOflag = F,
  plotCurrentYear = F
)
}
\arguments{
\item{speciescode}{Integer indicating the species code.}

\item{sitecode}{Integer or a list indicating site codes.}

\item{period}{String indicating if LFMC values are shown by month or fortnight.}

\item{quantiles}{Vector indicating the quantiles to be displayed.}

\item{LFMCflag}{Logical indicating if manually identified outliers are excluded.}

\item{AOflag}{Logical indicating if additive outliers are excluded.}

\item{plotCurrentYear}{Logical indicating if data from the last two years are shown.}
}
\value{
A seasonal plot by site
}
\description{
Seasonal plots with quantiles for LFMC data
}
\examples{
\dontrun{
   seasonalPlot(speciescode = 1, sitecode = c(1, 5),
   period = "Month", quantiles = c(0.1, 0.25, 0.50, 0.75, 0.97), data = T)
}

}
