% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setDBpath.R
\name{setDBpath}
\alias{setDBpath}
\title{Set the LFMC database file}
\usage{
setDBpath(file)
}
\arguments{
\item{file}{A string with the path to a *.sqlite file}
}
\description{
Specifies the location of the LFMC database file. This causes setting global variable \code{lfmcdbfile}.
}
\examples{
\dontrun{
   setDBpath("../data/lfmc.sqlite")
}
}
